// ebk_front/home/home/app/lib/config/interface.ts
export interface Users {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  nom: string;
  prenom: string;
  telephone: string;
  email: string;
  sexe: null | string;
  profil: null | string;
  datenaissance: Date | null;
  adresse: null | string;
  ville: null | string;
  pays: null | string;
  username: null;
  salt: string;
  password: string;
  hashedRt: string;
  privilege: PrivilegesEnum;
  status: string;
  confirm: boolean;
  eglise: Eglise;
}
export interface Eglise {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id_eglise: number;
  nom_eglise: string;
  username_eglise: string;
  photo_eglise: string;
  couverture_eglise: string;
  sigle_eglise: string;
  adresse_eglise: string;
  ville_eglise: string;
  pays_eglise: string;
  nombrefidel_eglise: string;
  status_eglise: string;
  payement_eglise: boolean;
  programme?: Programme[];
  word_of_day_eglise: string;
  about_eglise: string;
  localisation_eglise: string[];
}
// ...
export interface Commentaires {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  commentaire: string;
  users: Users;
}
export interface Favoris {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  users?: Users;
}

export interface Likes {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  users?: Users;
}
export interface Programme {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  sousProgramme: SousProgramme[];
}
export interface SousProgramme {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  debut: Date;
  fin: Date;
  libelle: string;
}
export interface Communiques {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  id: number;
  communique: string;
}
export interface subjectForum {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  id: number;
  title: string;
  description: string;
}
// end ........ interface for get or find data  ...... end

// interface for get or find data paginated
export interface ItemPicture {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date | null;
  id: number;
  descrition: string;
  photos: string[];
  sharecode: null | string;
  eglise: Eglise;
  favoris: Favoris[];
  commentaire: Commentaires[];
  share: any[];
  likes: Likes[];
}
export interface ItemVideos {
  views: any;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  lien: string;
  photo: string;
  webp: null;
  auteur: string;
  interne: boolean;
  sharecode: string;
  eglise: Eglise;
  favoris: Favoris[];
  likes: Likes[];
  commentaire: Commentaires[];
  share: Favoris[];
}
// ...
export interface ItemBibleStudy {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  objectif: string;
  description: string;
  payement: string;
  sharecode: string;
  favoris: Favoris[];
  commentaire: Commentaires[];
  share: any[];
  likes: Likes[];
  contentsBibleStudy: ItemContentBibleStudy[];
  eglise: Eglise;
}

export interface ItemContentBibleStudy {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  image: string;
  content: string;
  sharecode: string;
  favoris: Favoris[];
  commentaire: Commentaires[];
  share: any[];
  likes: Likes[];
}
// ...
export interface SearchIterface {
  images: ItemPicture[];
  audios: ItemVideos[];
  videos: ItemVideos[];
  live: ItemVideos[];
  livre: ItemVideos[];
  eglise: Eglise[];
  bibleStudy: ItemBibleStudy[];
  forum: any[];
}